### 2020-08-27 #2605 @whedon
[REVIEW]: NO-DOUBT: A Decision Support Web Platform based on AHP
**Submitting author:** @gdede-hua (<a href="http://orcid.org/0000-0003-2947-7470">Georgia Dede</a>)
**Repository:** <a href="https://github.com/gdede-hua/decision-survey-platform" target ="_blank">https://github.com/gdede-hua/decision-survey-platform</a>
**Version:** v1.0
