### 2020-08-27 #2605 @remram44
It seems that the paper is missing figures. In addition the references don't render correctly, because the `title` of most entries are URL, which get capitalized (`Https://github.com/airicyu/ahp`).

I assume the paper is still being worked on but I thought I'd make a note of it regardless.
