### 2020-08-28 #2605 @gdede-hua
@sbenthall <https://github.com/sbenthall> I have checked the paper and it
worked. It also passed the Joss compilation. My apologies. Do you want to
revise the paper and resubmit it?

On Fri, Aug 28, 2020 at 2:17 AM Sebastian Benthall <notifications@github.com>
wrote:

> Does that sound right, @gdede-hua <https://github.com/gdede-hua>?
>
