### 2020-08-31 #2605 @gdede-hua
@sbenthall I have received the proofs of my paper. May I kindly ask you to
inform how I can include the figures in the paper?
