### 2020-09-01 #2605 @alexssandrofc
Hi @gdede-hua,

I hope to help you, I made available some tips for the development of your paper.

Figures can be included like this:
![Caption for example figure.\label{fig:example}](figure.png)

As an example available in https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

It is possible to check a preview of your paper at https://whedon.theoj.org/

I also verified that the URLs of the images in your paper.md are incorrect, example of same image i believe:

https://github.com/gdede-hua/decision-survey-platform/blob/master/create_alternatives -> URL in paper.md with error 404.
https://github.com/alexandrakisdimi/decision-making-system/blob/master/screenshots/create_alternatives.png  ->  URL ok in README.md
