### 2020-09-11 #2605 @gdede-hua
@alexssandrofc  I have made the modifications in my paper according to your instructions, but the whedon paper preview service appears this message:

"Your paper failed to compile with the following errors:

{}"

It seems there is no error but the paper is not compiled. May I kindly ask you to further assist me on this issue?

Thank you.
