### 2020-09-14 #2605 @alexssandrofc
@gdede-hua as I checked your paper.md, I believe the error might be on this line:

-> [User Answer Example 4.\label{fig:user_answer4](screenshots/user_answer4.png)

"}" is missing to close the (\label).
