### 2020-09-14 #2605 @remram44
The `{}` part seems like a bug. I'm not sure if it should be reported against [whedon](https://github.com/openjournals/whedon) or [whedon-api](https://github.com/openjournals/whedon-api), but it certainly makes it hard to fix papers...
