### 2020-09-14 #2605 @alexssandrofc
the problem was in the API (https://whedon.theoj.org/), when executing the command to generate the PDF, it generated correctly.
