### 2020-09-16 #2605 @gdede-hua
@alexssandrofc Thank you very much. I checked the article proof. It's correct.

@sbenthall Thank you for your support. The article proof is correct. So if you wish we are ready to proceed with the review.
