### 2020-10-05 #2605 @remram44
Apologies for the delay, I am reviewing today.

Some notes before I go into running the software:

* The commit history is very short (June 28 - July 9)
  * @gdede-hua is the only committer, along with one macro-commit from Dimitrios Alexandrakis [`39945023`](https://github.com/gdede-hua/decision-survey-platform/commit/39945023795e64f4b42146ff410379291f3fd6ef). Hard to tell if the author list is correct from Git data (but it's probably fine if @gdede-hua says it is)
  * The commits don't seem to reflect the development history, huge chunks of code were imported in one go. The JOSS guidelines require "not less that three months of work" but I don't think we should judge this from the commit history in this instance
  * About 6,000 LoC in total excluding vendored dependencies
* Reviewed the paper
  * Some minor fixes submitted as a PR: https://github.com/gdede-hua/decision-survey-platform/pull/3 (mostly grammar, punctuation, capitalizing tool and company names)
  * The paper claims that many implementations of the algorithm exist, but this is the only AHP tool that offers a web-based interface for both administrators and users, which I think is a sufficient statement of need
  * "PWC" should be defined
  * One paragraph ends abruptly, it might just be missing a period but maybe that sentence was meant to say more
  * "materialized algorithm" seemed odd, but it might just be me
