### 2020-10-05 #2605 @remram44
I am still getting LAPACK etc set up, but going over the code:

* There are no tests
* Only PHP 5.6 is supported, which is end-of-life since 2018 (PHP 7 was released 2015). This is sure to be a problem going forward
* Some content seems to still be in Greek (mostly button labels, some `alert()` calls)
* There is significant code duplication
  * [`en/admin/quest3.php`](https://github.com/gdede-hua/decision-survey-platform/blob/264399f05aaaf62252f5d4dd86b7e62d1991d788/en/admin/quest3.php) and [`en/admin/quest_criteria2.php`](https://github.com/gdede-hua/decision-survey-platform/blob/264399f05aaaf62252f5d4dd86b7e62d1991d788/en/admin/quest_criteria2.php) are the same files entirely
  * other files share a lot of code, e.g. forms, rendering code. Consider using include or functions to make maintenance possible
  * some duplication even within single files, e.g. `en/admin/generate_results2.php`: lines [144-204](https://github.com/gdede-hua/decision-survey-platform/blob/264399f05aaaf62252f5d4dd86b7e62d1991d788/en/admin/generate_results2.php#L144-L205) and [265-328](https://github.com/gdede-hua/decision-survey-platform/blob/264399f05aaaf62252f5d4dd86b7e62d1991d788/en/admin/generate_results2.php#L265-L327) are the same
* The flow is also difficult to follow because of naming, e.g. there is a lot of files with similar names `quest2.php`, `quest2db.php`, `quest3.php`, ...
* There seems to be lost files like [this empty "test" file](https://github.com/gdede-hua/decision-survey-platform/blob/264399f05aaaf62252f5d4dd86b7e62d1991d788/en/admin/test) and some result spreadsheets: [`Car_buying_decision_Results.xlsx`](https://github.com/gdede-hua/decision-survey-platform/blob/264399f05aaaf62252f5d4dd86b7e62d1991d788/en/admin/Car_buying_decision_Results.xlsx), [`Choose_the_best_car_for_the_family_Results.xlsx`](https://github.com/gdede-hua/decision-survey-platform/blob/264399f05aaaf62252f5d4dd86b7e62d1991d788/en/admin/Choose_the_best_car_for_the_family_Results.xlsx)
* PHP, HTML, and JS code is mixed in single files, which partly causes the duplication described above. Referencing static JS files from `<script>` tags and using of HTML templating would make the code more readable and maintainable and would help with duplication (this is considered best practice)
