### 2020-10-05 #2605 @remram44
I managed to build LAPACK, install PHP 5.6, and run the software.

* Minor: The name of the system is wrong in the README (decision-making-system)
* The apache configuration file is missing (instructions say to "copy the `apache/decision-making-system.conf` into `/etc/apache2/sites-available/`" however the file is not in the repository). I created my own and was able to proceed
* MD5 is a very weak hashing algorithm and is never appropriate for securing passwords
* Most content is in English but some is entirely in Greek, like the "edit research" page
  * <details><summary>Screenshot</summary><img src="https://user-images.githubusercontent.com/426784/95139186-e5828a80-0739-11eb-9856-bcbf0a59d598.png"/></details>
  * This was probably easy to miss because of the code duplication, e.g. when you translated the "create research" page you left the duplicated "edit research" in Greek

I was able to create a research (I did it wrong and had to start over because it seems you can't change criteria and factors from the edit page), however I could not assign a user to it: <details><summary>"Setting users failed. Error 1"</summary><img src="https://user-images.githubusercontent.com/426784/95139520-a9035e80-073a-11eb-8414-b6e21de24580.png"></details>

Sorry again for the delay!
