### 2020-10-21 #2605 @alexssandrofc
First sorry for the delay in review,

Study with a very interesting topic, as it provides software to help in decision making on an easily accessible platform "WEB", so I would like to leave some contributions.

Below I leave some suggestions regarding the latex document:

-Topic "Statement Of Need", the point is missing at the end of the second paragraph, is there a continuation?

-PWC matrix in "AHP Methodology" has not been defined.

-In the "Sumary", the third paragraph contains the excerpt "materialized algorithms", I recommend changing only for "algorithms".

-Many images in a row without explanation.
