### 2020-10-21 #2605 @alexssandrofc
About the technologies used, only the alert when using PHP 5.6 is left, which is out of date and no longer receives official protection updates since January of 2019, making the site vulnerable to cyber attacks.

If possible, I recommend replacing the LAPACK library with another library that meets the needs of AHP, with compatibility with new versions of PHP.
