### 2020-10-21 #2605 @alexssandrofc
About the usability of the software:

-In the installation, "decision-making-system.conf" does not exist!
-When creating a user logged in as Administrator, the user created does not log in, however when creating a user from the login screen -> register, the registration of the new user occurs correctly.
-Recurring error when creating a survey, press the submit button and nothing happens without an error message and does not save the survey.
-When editing a survey, the page changes language in the middle of the process, changing from English to Greek.
-When completing the search and requesting the result, an "Not Found" error occurs and does not display the "xlsx".

I recommend a new round of tests to detect possible problems on the software .
